requirements.txt
nicegui
fastapi
uvicorn
openai
python-dotenv